pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-analyzer"
version = "0.1.0"
description = "Unified Streamlit + CLI financial analysis toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit",
  "langchain",
  "langgraph",
  "langchain-google-genai",
  "requests",
  "pandas",
  "yfinance",
  "yahooquery",
  "tavily-python",
  "numpy",
  "matplotlib",
  "beautifulsoup4",
  "Pillow",
  "fpdf",
  "deap",
  "PyPortfolioOpt",
  "xlrd>=2.0.1",
]

[project.scripts]
portfolio-analyzer = "portfolio_analyzer.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["E402"]
